Refactor code
Changes
1. Update _coord_desc.pyx to have specific integer types

2. Add compare_experiments. This is a script which will allow us
to quickly compute metrics from experiments and do so in a general
way so that we can quickly run multiple benchmarks analyses

3. Changed the expected format for the label map from a list of lists
to a NumPy array. This is easier and allows us to do some of our
other tasks quicker

4. Adjusted the hc.py to ensure that we're using the same validation
set throughout versus recycling multiple validation sets. Additionally
we're also tracking the `good_idx` so that if we have an issue
at the leaf level with a HC, we can pass that along to the metrics
we're computing

5. Add leaf_stability. This allows us to compute entropy at the leaf
level and do some analysis in that direction.

6. Add run_model. This is a general script that will allow us run a
particular model and save the results to disk so that we easily
execute multiple analyses without having to re-write similar code.

7. Add get_leaf_stability and test_hc as scripts which call CORE
for the FMoW experiments.

+++ core/_coord_desc.pyx
@@ -160,8 +160,8 @@ cpdef _run_search(np.ndarray[DOUBLE, ndim=2] V, np.ndarray[INT, ndim=1] z,
         DOUBLE obj_val_old
         int c = V.shape[0]
         int l = mu.shape[0]
-        np.ndarray[INT, ndim=1] metaclasses = np.arange(l)
-        np.ndarray[INT, ndim=1] labels = np.arange(c)
+        np.ndarray[INT, ndim=1] metaclasses = np.arange(l, dtype=np.int32)
+        np.ndarray[INT, ndim=1] labels = np.arange(c, dtype=np.int32)
         np.ndarray[INT, ndim=1] label_options
         int iter_num
         int i
